refactor: Move block prevalidation logic to core
