+++ node/src/state.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 
 use malloc_size_of_derive::MallocSizeOf;
 use mina_p2p_messages::v2;
-use openmina_core::constants::PROTOCOL_VERSION;
+use openmina_core::block::prevalidate::{prevalidate_block, BlockPrevalidationError};
 use openmina_core::transaction::{TransactionInfo, TransactionWithHash};
 use p2p::P2pNetworkPubsubMessageCacheId;
 use rand::prelude::*;
@@ -78,25 +78,6 @@ pub struct State {
     // TODO(binier): include action kind in `last_action`.
     last_action: ActionMeta,
     applied_actions_count: u64,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub enum BlockPrevalidationError {
-    GenesisNotReady,
-    ReceivedTooEarly {
-        current_global_slot: u32,
-        block_global_slot: u32,
-    },
-    ReceivedTooLate {
-        current_global_slot: u32,
-        block_global_slot: u32,
-        delta: u32,
-    },
-    InvalidGenesisProtocolState,
-    InvalidProtocolVersion,
-    MismatchedProtocolVersion,
-    ConsantsMismatch,
-    InvalidDeltaBlockChainProof,
 }
 
 // Substate accessors that will be used in reducers
@@ -408,70 +389,7 @@ impl State {
             return Err(BlockPrevalidationError::GenesisNotReady);
         };
 
-        // received_at_valid_time
-        // https://github.com/minaprotocol/mina/blob/6af211ad58e9356f00ea4a636cea70aa8267c072/src/lib/consensus/proof_of_stake.ml#L2746
-        {
-            let block_global_slot = block.global_slot();
-
-            let delta = genesis.constants().delta.as_u32();
-            if cur_global_slot < block_global_slot {
-                // Too_early
-                return Err(BlockPrevalidationError::ReceivedTooEarly {
-                    current_global_slot: cur_global_slot,
-                    block_global_slot,
-                });
-            } else if !allow_block_too_late
-                && cur_global_slot.saturating_sub(block_global_slot) > delta
-            {
-                // Too_late
-                return Err(BlockPrevalidationError::ReceivedTooLate {
-                    current_global_slot: cur_global_slot,
-                    block_global_slot,
-                    delta,
-                });
-            }
-        }
-
-        if block.header().genesis_state_hash() != genesis.hash() {
-            return Err(BlockPrevalidationError::InvalidGenesisProtocolState);
-        }
-
-        let (protocol_versions_are_valid, protocol_version_matches_daemon) = {
-            let min_transaction_version = 1.into();
-            let v = &block.header().current_protocol_version;
-            let nv = block
-                .header()
-                .proposed_protocol_version_opt
-                .as_ref()
-                .unwrap_or(v);
-
-            // Our version values are unsigned, so there is no need to check that the
-            // other parts are not negative.
-            let valid = v.transaction >= min_transaction_version
-                && nv.transaction >= min_transaction_version;
-            let compatible = v.transaction == PROTOCOL_VERSION.transaction
-                && v.network == PROTOCOL_VERSION.network;
-
-            (valid, compatible)
-        };
-
-        if !protocol_versions_are_valid {
-            return Err(BlockPrevalidationError::InvalidProtocolVersion);
-        } else if !protocol_version_matches_daemon {
-            return Err(BlockPrevalidationError::MismatchedProtocolVersion);
-        }
-
-        // NOTE: currently these cannot change between blocks, but that
-        // may not always be true?
-        if block.constants() != genesis.constants() {
-            return Err(BlockPrevalidationError::ConsantsMismatch);
-        }
-
-        // TODO(tizoc): check for InvalidDeltaBlockChainProof
-        // https://github.com/MinaProtocol/mina/blob/d800da86a764d8d37ffb8964dd8d54d9f522b358/src/lib/mina_block/validation.ml#L369
-        // https://github.com/MinaProtocol/mina/blob/d800da86a764d8d37ffb8964dd8d54d9f522b358/src/lib/transition_chain_verifier/transition_chain_verifier.ml
-
-        Ok(())
+        prevalidate_block(block, &genesis, cur_global_slot, allow_block_too_late)
     }
 
     pub fn should_log_node_id(&self) -> bool {
